Revert "Move Tag out of their own package."
This reverts commit b91ee5545b8353da9b7a9f14ce242ab5874a72e3.
